test: move process lifecycle tests to their own file for modularity
